+++ rigidBody.hpp
@@ -8,52 +8,52 @@ using Clock = std::chrono::high_resolution_clock;
 using TimePoint = std::chrono::time_point<Clock>;
 using Duration = std::chrono::duration<double>;
 
-extern float springConstant; // N/m
+extern double springConstant; // N/m
 
 class RigidBody {
     private:
-        float radius_; // m
-        float mass_; // kg
-        float invMass_; // 1/kg (inverse mass for physics calculations)
-        std::array<float, 2> position_; // m
-        std::array<float, 2> velocity_; // m/s
-        std::array<float, 2> acceleration_; // m/s^2
-        std::array<float, 2> force_; // N
+        double radius_; // m
+        double mass_; // kg
+        double invMass_; // 1/kg (inverse mass for physics calculations)
+        std::array<double, 2> position_; // m
+        std::array<double, 2> velocity_; // m/s
+        std::array<double, 2> acceleration_; // m/s^2
+        std::array<double, 2> force_; // N
 
     public:
         // Constructor
-        RigidBody(float radius = 1.0f, float mass = 1.0f, 
-                  std::array<float, 2> position = {0.0f, 0.0f},
-                  std::array<float, 2> velocity = {0.0f, 0.0f});
+        RigidBody(double radius = 1.0, double mass = 1.0, 
+                  std::array<double, 2> position = {0.0, 0.0},
+                  std::array<double, 2> velocity = {0.0, 0.0});
         // Getters
-        float getRadius() const { return radius_; }
-        float getMass() const { return mass_; }
-        float getInvMass() const { return invMass_; }
-        std::array<float, 2> getPosition() const { return position_; }
-        std::array<float, 2> getVelocity() const { return velocity_; }
-        std::array<float, 2> getAcceleration() const { return acceleration_; }
-        std::array<float, 2> getForce() const { return force_; }
+        double getRadius() const { return radius_; }
+        double getMass() const { return mass_; }
+        double getInvMass() const { return invMass_; }
+        std::array<double, 2> getPosition() const { return position_; }
+        std::array<double, 2> getVelocity() const { return velocity_; }
+        std::array<double, 2> getAcceleration() const { return acceleration_; }
+        std::array<double, 2> getForce() const { return force_; }
 
         // Setters
-        void setRadius(float radius) { radius_ = radius; }
-        void setMass(float mass) { 
+        void setRadius(double radius) { radius_ = radius; }
+        void setMass(double mass) { 
             mass_ = mass; 
-            invMass_ = (mass > 0.0f) ? (1.0f / mass) : 0.0f;
+            invMass_ = (mass > 0.0) ? (1.0 / mass) : 0.0;
         }
-        void setPosition(const std::array<float, 2>& position) { position_ = position; }
-        void setVelocity(const std::array<float, 2>& velocity) { velocity_ = velocity; }
-        void setAcceleration(const std::array<float, 2>& acceleration) { acceleration_ = acceleration; }
-        void setForce(const std::array<float, 2>& force) { force_ = force; }
+        void setPosition(const std::array<double, 2>& position) { position_ = position; }
+        void setVelocity(const std::array<double, 2>& velocity) { velocity_ = velocity; }
+        void setAcceleration(const std::array<double, 2>& acceleration) { acceleration_ = acceleration; }
+        void setForce(const std::array<double, 2>& force) { force_ = force; }
 
         // Individual component setters cuz we lazy
-        void setPositionX(float x) { position_[0] = x; }
-        void setPositionY(float y) { position_[1] = y; }
-        void setVelocityX(float vx) { velocity_[0] = vx; }
-        void setVelocityY(float vy) { velocity_[1] = vy; }
-        void setForceX(float fx) { force_[0] = fx; }
-        void setForceY(float fy) { force_[1] = fy; }
+        void setPositionX(double x) { position_[0] = x; }
+        void setPositionY(double y) { position_[1] = y; }
+        void setVelocityX(double vx) { velocity_[0] = vx; }
+        void setVelocityY(double vy) { velocity_[1] = vy; }
+        void setForceX(double fx) { force_[0] = fx; }
+        void setForceY(double fy) { force_[1] = fy; }
 
-        void numericalIntegration(float dt);
+        void numericalIntegration(double dt);
 };
 
 class TimeManager {
